requirements.txt
# --- Core Flask Stack ---
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0

# --- Utilities ---
python-dotenv==1.1.1
colorama==0.4.6
typing_extensions==4.15.0

# --- Database & ORM ---
SQLAlchemy==2.0.43
greenlet==3.2.4
alembic==1.16.5
Mako==1.3.10
MarkupSafe==3.0.3

# --- Machine Learning / Vision ---
numpy==1.26.4
tensorflow==2.16.1
tf-keras==2.17.0
deepface==0.0.93

# --- Development ---
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
